refactoring: user avatar & custom hook
UserAvatar 폴더로 이동
아이콘을 wrapping 된 컴포넌트로 교체
