+++ frontend/guest-app/src/components/UserAvatar/useUserAvatar.js
